agent: Rename utils/file_utils.py to utils/dir_utils.py
